requirements.txt
pdfplumber
python-dotenv
openai
pydantic
psutil
pyyaml
requests
